09_fine-turnning/unsloth/requirements.txt
unsloth
wandb
python-dotenv
vllm
deepspeed